requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
PyGithub
atlassian-python-api
python-dotenv
jira
requests
pydantic
